Added querying and removing plugins.
